New algorithms: accumulate{,_xform}{,_if}{,_n}
This accumulate() uses
   acc += *first
instead of
   acc = acc + *first,
making it much more efficient for, say, strings.

Consequently, the binary function passed is not an operation,
like in std::accumulate(), but an action (ie. it mutates its
first argument, like +=).

Function objects to go with this algorithm are still missing
(plus_equal, multiply_equal? Or add_to, multiply_to?)